+++ modules/FormS.py
@@ -1,0 +1,260 @@
+# -*- coding: utf-8 -*-
+
+################################################################################
+## Form generated from reading UI file 'FormWiZPBX.ui'
+##
+## Created by: Qt User Interface Compiler version 5.15.2
+##
+## WARNING! All changes made in this file will be lost when recompiling UI file!
+################################################################################
+
+from PySide2.QtCore import *
+from PySide2.QtGui import *
+from PySide2.QtWidgets import *
+
+
+class Ui_Auxprog(object):
+    def setupUi(self, Auxprog):
+        if not Auxprog.objectName():
+            Auxprog.setObjectName(u"Auxprog")
+        Auxprog.resize(720, 471)
+        icon = QIcon()
+        icon.addFile(u"../../../../Pictures/Saved Pictures/\u041f\u0443\u0442\u0438\u043d.jpg", QSize(), QIcon.Normal, QIcon.Off)
+        Auxprog.setWindowIcon(icon)
+        Auxprog.setStyleSheet(u".QPushButton {\n"
+"\n"
+"}\n"
+"#PushButton {\n"
+"}")
+        self.graphicsView = QGraphicsView(Auxprog)
+        self.graphicsView.setObjectName(u"graphicsView")
+        self.graphicsView.setEnabled(True)
+        self.graphicsView.setGeometry(QRect(10, 40, 691, 192))
+        self.graphicsView.setStyleSheet(u".QGraphicsView {\n"
+"	\n"
+"	background-color: rgba(255, 255, 255, 125);\n"
+"}")
+        self.graphicsView.setInteractive(True)
+        self.graphicsView.setSceneRect(QRectF(0, 0, 0, 0))
+        self.Startstop = QPushButton(Auxprog)
+        self.Startstop.setObjectName(u"Startstop")
+        self.Startstop.setGeometry(QRect(30, 140, 131, 41))
+        self.label = QLabel(Auxprog)
+        self.label.setObjectName(u"label")
+        self.label.setGeometry(QRect(20, 0, 81, 41))
+        font = QFont()
+        font.setPointSize(18)
+        self.label.setFont(font)
+        self.label_2 = QLabel(Auxprog)
+        self.label_2.setObjectName(u"label_2")
+        self.label_2.setGeometry(QRect(20, 50, 91, 41))
+        font1 = QFont()
+        font1.setPointSize(10)
+        self.label_2.setFont(font1)
+        self.label_3 = QLabel(Auxprog)
+        self.label_3.setObjectName(u"label_3")
+        self.label_3.setGeometry(QRect(220, 50, 91, 41))
+        self.label_3.setFont(font1)
+        self.radioButton = QRadioButton(Auxprog)
+        self.radioButton.setObjectName(u"radioButton")
+        self.radioButton.setGeometry(QRect(320, 60, 82, 31))
+        self.radioButton_2 = QRadioButton(Auxprog)
+        self.radioButton_2.setObjectName(u"radioButton_2")
+        self.radioButton_2.setGeometry(QRect(320, 80, 82, 31))
+        self.label_4 = QLabel(Auxprog)
+        self.label_4.setObjectName(u"label_4")
+        self.label_4.setGeometry(QRect(370, 50, 81, 41))
+        self.label_4.setFont(font1)
+        self.checkBox = QCheckBox(Auxprog)
+        self.checkBox.setObjectName(u"checkBox")
+        self.checkBox.setGeometry(QRect(430, 60, 70, 17))
+        self.checkBox_2 = QCheckBox(Auxprog)
+        self.checkBox_2.setObjectName(u"checkBox_2")
+        self.checkBox_2.setGeometry(QRect(430, 90, 70, 17))
+        self.checkBox_3 = QCheckBox(Auxprog)
+        self.checkBox_3.setObjectName(u"checkBox_3")
+        self.checkBox_3.setGeometry(QRect(430, 120, 111, 17))
+        self.checkBox_4 = QCheckBox(Auxprog)
+        self.checkBox_4.setObjectName(u"checkBox_4")
+        self.checkBox_4.setGeometry(QRect(430, 150, 161, 16))
+        self.checkBox_5 = QCheckBox(Auxprog)
+        self.checkBox_5.setObjectName(u"checkBox_5")
+        self.checkBox_5.setGeometry(QRect(430, 180, 131, 17))
+        self.label_freze = QLabel(Auxprog)
+        self.label_freze.setObjectName(u"label_freze")
+        self.label_freze.setGeometry(QRect(230, 130, 111, 41))
+        self.label_freze.setFont(font1)
+        self.pushButton_2 = QPushButton(Auxprog)
+        self.pushButton_2.setObjectName(u"pushButton_2")
+        self.pushButton_2.setGeometry(QRect(270, 170, 75, 23))
+        self.pushButton_3 = QPushButton(Auxprog)
+        self.pushButton_3.setObjectName(u"pushButton_3")
+        self.pushButton_3.setGeometry(QRect(270, 200, 75, 23))
+        self.label_6 = QLabel(Auxprog)
+        self.label_6.setObjectName(u"label_6")
+        self.label_6.setGeometry(QRect(20, 230, 81, 41))
+        self.label_6.setFont(font)
+        self.graphicsView_2 = QGraphicsView(Auxprog)
+        self.graphicsView_2.setObjectName(u"graphicsView_2")
+        self.graphicsView_2.setEnabled(True)
+        self.graphicsView_2.setGeometry(QRect(10, 270, 691, 192))
+        self.graphicsView_2.setStyleSheet(u".QGraphicsView {\n"
+"	\n"
+"	background-color: rgba(255, 255, 255, 125);\n"
+"}")
+        self.graphicsView_2.setInteractive(True)
+        self.graphicsView_2.setSceneRect(QRectF(0, 0, 0, 0))
+        self.pushButton_4 = QPushButton(Auxprog)
+        self.pushButton_4.setObjectName(u"pushButton_4")
+        self.pushButton_4.setGeometry(QRect(30, 400, 131, 41))
+        self.label_7 = QLabel(Auxprog)
+        self.label_7.setObjectName(u"label_7")
+        self.label_7.setGeometry(QRect(30, 300, 101, 41))
+        self.label_7.setFont(font1)
+        self.pushButton_5 = QPushButton(Auxprog)
+        self.pushButton_5.setObjectName(u"pushButton_5")
+        self.pushButton_5.setGeometry(QRect(140, 310, 61, 23))
+        self.pushButton_6 = QPushButton(Auxprog)
+        self.pushButton_6.setObjectName(u"pushButton_6")
+        self.pushButton_6.setGeometry(QRect(170, 400, 151, 41))
+        self.label_8 = QLabel(Auxprog)
+        self.label_8.setObjectName(u"label_8")
+        self.label_8.setGeometry(QRect(530, 260, 101, 41))
+        self.label_8.setFont(font1)
+        self.pushButton_7 = QPushButton(Auxprog)
+        self.pushButton_7.setObjectName(u"pushButton_7")
+        self.pushButton_7.setGeometry(QRect(440, 300, 251, 23))
+        font2 = QFont()
+        font2.setBold(True)
+        font2.setUnderline(True)
+        font2.setWeight(75)
+        self.pushButton_7.setFont(font2)
+        self.pushButton_8 = QPushButton(Auxprog)
+        self.pushButton_8.setObjectName(u"pushButton_8")
+        self.pushButton_8.setGeometry(QRect(440, 330, 251, 23))
+        self.pushButton_8.setFont(font2)
+        self.pushButton_9 = QPushButton(Auxprog)
+        self.pushButton_9.setObjectName(u"pushButton_9")
+        self.pushButton_9.setGeometry(QRect(440, 360, 251, 23))
+        self.pushButton_9.setFont(font2)
+        self.pushButton_10 = QPushButton(Auxprog)
+        self.pushButton_10.setObjectName(u"pushButton_10")
+        self.pushButton_10.setGeometry(QRect(440, 390, 251, 23))
+        self.pushButton_10.setFont(font2)
+        self.label_9 = QLabel(Auxprog)
+        self.label_9.setObjectName(u"label_9")
+        self.label_9.setGeometry(QRect(330, 260, 111, 41))
+        self.label_9.setFont(font1)
+        self.pushButton_11 = QPushButton(Auxprog)
+        self.pushButton_11.setObjectName(u"pushButton_11")
+        self.pushButton_11.setGeometry(QRect(350, 310, 61, 23))
+        self.graphicsView_3 = QGraphicsView(Auxprog)
+        self.graphicsView_3.setObjectName(u"graphicsView_3")
+        self.graphicsView_3.setEnabled(True)
+        self.graphicsView_3.setGeometry(QRect(0, 0, 721, 471))
+        self.graphicsView_3.setStyleSheet(u".QGraphicsView {\n"
+"	\n"
+"	background-color: qlineargradient(spread:pad, x1:0, y1:1, x2:0, y2:0, stop:0 rgba(0, 0, 0, 255), stop:0.05 rgba(14, 8, 73, 255), stop:0.36 rgba(28, 17, 145, 255), stop:0.6 rgba(126, 14, 81, 255), stop:0.75 rgba(234, 11, 11, 255), stop:0.79 rgba(244, 70, 5, 255), stop:0.86 rgba(255, 136, 0, 255), stop:0.935 rgba(239, 236, 55, 255));\n"
+"}")
+        self.graphicsView_3.setInteractive(True)
+        self.graphicsView_3.setSceneRect(QRectF(0, 0, 0, 0))
+        self.horizontalSlider = QSlider(Auxprog)
+        self.horizontalSlider.setObjectName(u"horizontalSlider")
+        self.horizontalSlider.setGeometry(QRect(200, 360, 121, 22))
+        self.horizontalSlider.setOrientation(Qt.Horizontal)
+        self.label_10 = QLabel(Auxprog)
+        self.label_10.setObjectName(u"label_10")
+        self.label_10.setGeometry(QRect(30, 350, 171, 41))
+        self.label_10.setFont(font1)
+        self.posX = QLabel(Auxprog)
+        self.posX.setObjectName(u"posX")
+        self.posX.setGeometry(QRect(630, 90, 47, 16))
+        self.posY = QLabel(Auxprog)
+        self.posY.setObjectName(u"posY")
+        self.posY.setGeometry(QRect(630, 110, 47, 13))
+        self.posY_2 = QLabel(Auxprog)
+        self.posY_2.setObjectName(u"posY_2")
+        self.posY_2.setGeometry(QRect(550, 40, 141, 41))
+        self.Delay01 = QLineEdit(Auxprog)
+        self.Delay01.setObjectName(u"Delay01")
+        self.Delay01.setGeometry(QRect(110, 60, 51, 20))
+        self.graphicsView_3.raise_()
+        self.graphicsView.raise_()
+        self.Startstop.raise_()
+        self.label.raise_()
+        self.label_2.raise_()
+        self.label_3.raise_()
+        self.radioButton.raise_()
+        self.radioButton_2.raise_()
+        self.label_4.raise_()
+        self.checkBox.raise_()
+        self.checkBox_2.raise_()
+        self.checkBox_3.raise_()
+        self.checkBox_4.raise_()
+        self.checkBox_5.raise_()
+        self.label_freze.raise_()
+        self.pushButton_2.raise_()
+        self.pushButton_3.raise_()
+        self.label_6.raise_()
+        self.graphicsView_2.raise_()
+        self.pushButton_4.raise_()
+        self.label_7.raise_()
+        self.pushButton_5.raise_()
+        self.pushButton_6.raise_()
+        self.label_8.raise_()
+        self.pushButton_7.raise_()
+        self.pushButton_8.raise_()
+        self.pushButton_9.raise_()
+        self.pushButton_10.raise_()
+        self.label_9.raise_()
+        self.pushButton_11.raise_()
+        self.horizontalSlider.raise_()
+        self.label_10.raise_()
+        self.posX.raise_()
+        self.posY.raise_()
+        self.posY_2.raise_()
+        self.Delay01.raise_()
+
+        self.retranslateUi(Auxprog)
+        self.pushButton_4.clicked.connect(self.horizontalSlider.hide)
+        self.pushButton_3.clicked.connect(self.label_freze.update)
+
+        QMetaObject.connectSlotsByName(Auxprog)
+    # setupUi
+
+    def retranslateUi(self, Auxprog):
+        Auxprog.setWindowTitle(QCoreApplication.translate("Auxprog", u"Auxprog", None))
+        self.Startstop.setText(QCoreApplication.translate("Auxprog", u"\u0421\u0442\u0430\u0440\u0442/\u0421\u0442\u043e\u043f", None))
+        self.label.setText(QCoreApplication.translate("Auxprog", u"\u041a\u043b\u0438\u043a\u0435\u0440", None))
+        self.label_2.setText(QCoreApplication.translate("Auxprog", u"\u0417\u0430\u0434\u0435\u0440\u0436\u043a\u0430(\u043c\u0441)", None))
+        self.label_3.setText(QCoreApplication.translate("Auxprog", u"\u041a\u043e\u043b-\u0432\u043e \u043a\u043b\u0438\u043a\u043e\u0432", None))
+        self.radioButton.setText(QCoreApplication.translate("Auxprog", u"1", None))
+        self.radioButton_2.setText(QCoreApplication.translate("Auxprog", u"2", None))
+        self.label_4.setText(QCoreApplication.translate("Auxprog", u"\u0414\u0440\u0443\u0433\u043e\u0435", None))
+        self.checkBox.setText(QCoreApplication.translate("Auxprog", u"\u041b\u041a\u041c", None))
+        self.checkBox_2.setText(QCoreApplication.translate("Auxprog", u"\u041f\u041a\u041c", None))
+        self.checkBox_3.setText(QCoreApplication.translate("Auxprog", u"\u041f\u0435\u0440\u0435\u0440\u044b\u0432(5\u043c\u0438\u043d\u0443\u0442)", None))
+        self.checkBox_4.setText(QCoreApplication.translate("Auxprog", u"\u041f\u043e\u0441\u0442\u043e\u043e\u0431\u0440\u043e\u0432\u043d\u0430\u044f \u043e\u0431\u0440\u0430\u0431\u043e\u0442\u043a\u0430", None))
+        self.checkBox_5.setText(QCoreApplication.translate("Auxprog", u"\u0420\u0430\u0437\u0433\u043e\u043d(\u0417\u0433\u0430\u0436\u0438\u0432\u0430\u043d\u0438\u0435)", None))
+        self.label_freze.setText(QCoreApplication.translate("Auxprog", u"\u0417\u0430\u043c\u043e\u0440\u043e\u0437\u0438\u0442\u044c \u043a\u0443\u0440\u0441\u043e\u0440", None))
+        self.pushButton_2.setText(QCoreApplication.translate("Auxprog", u"\u0414\u0430", None))
+        self.pushButton_3.setText(QCoreApplication.translate("Auxprog", u"\u041d\u0435\u0442", None))
+        self.label_6.setText(QCoreApplication.translate("Auxprog", u"\u041c\u0430\u043a\u0440\u043e\u0441\u044b", None))
+        self.pushButton_4.setText(QCoreApplication.translate("Auxprog", u"\u0421\u0442\u0430\u0440\u0442/\u0421\u0442\u043e\u043f", None))
+        self.label_7.setText(QCoreApplication.translate("Auxprog", u"\u0412\u044b\u0431\u0440\u0430\u0442\u044c \u043c\u0430\u043a\u0440\u043e\u0441", None))
+        self.pushButton_5.setText(QCoreApplication.translate("Auxprog", u"\u041e\u0442\u043a\u0440\u044b\u0442\u044c", None))
+        self.pushButton_6.setText(QCoreApplication.translate("Auxprog", u"\u0421\u043e\u0437\u0434\u0430\u0442\u044c \u043c\u0430\u043a\u0440\u043e\u0441", None))
+        self.label_8.setText(QCoreApplication.translate("Auxprog", u"\u041e\u0431\u0443\u0447\u0435\u043d\u0438\u044f", None))
+        self.pushButton_7.setText(QCoreApplication.translate("Auxprog", u"\u041a\u0430\u043a \u043f\u043e\u043b\u044c\u0437\u043e\u0432\u0430\u0442\u044c\u0441\u044f \u043f\u0440\u043e\u0433\u0440\u0430\u043c\u043c\u043e\u0439", None))
+        self.pushButton_8.setText(QCoreApplication.translate("Auxprog", u"\u041a\u0430\u043a \u0441\u043e\u0437\u0434\u0430\u0442\u044c \u0441\u0432\u043e\u0439 \u043c\u0430\u043a\u0440\u043e\u0441", None))
+        self.pushButton_9.setText(QCoreApplication.translate("Auxprog", u"\u041a\u0430\u043a \u0437\u0430\u0433\u0440\u0443\u0437\u0438\u0442\u044c \u043c\u0430\u043a\u0440\u043e\u0441", None))
+        self.pushButton_10.setText(QCoreApplication.translate("Auxprog", u"\u041e\u0442\u043d\u043e\u0448\u0435\u043d\u0438\u0435 \u0434\u0440\u0443\u0433\u0438\u0445 \u043f\u0440\u043e\u0433\u0440\u0430\u043c\u043c \u043a \u043c\u0430\u043a\u0440\u043e\u0441\u0430\u043c", None))
+        self.label_9.setText(QCoreApplication.translate("Auxprog", u"\u0413\u043e\u0440\u044f\u0447\u0438\u0435 \u043a\u043b\u0430\u0432\u0438\u0448\u0438", None))
+        self.pushButton_11.setText(QCoreApplication.translate("Auxprog", u"+", None))
+        self.label_10.setText(QCoreApplication.translate("Auxprog", u"\u0421\u043a\u043e\u0440\u043e\u0441\u0442\u044c \u0432\u043e\u0438\u0441\u043f\u0440\u043e\u0438\u0437\u0432\u0435\u0434\u0435\u043d\u0438\u044f", None))
+        self.posX.setText(QCoreApplication.translate("Auxprog", u"0", None))
+        self.posY.setText(QCoreApplication.translate("Auxprog", u"0", None))
+        self.posY_2.setText(QCoreApplication.translate("Auxprog", u"<html><head/><body><p><span style=\" font-size:12pt;\">\u041f\u043e\u0437\u0438\u0446\u0438\u044f \u043a\u0443\u0440\u0441\u043e\u0440\u0430:</span></p></body></html>", None))
+        self.Delay01.setText(QCoreApplication.translate("Auxprog", u"100", None))
+    # retranslateUi
+
