+++ modules/Main.py
@@ -1,8 +1,8 @@
 import pyautogui as a
 from PySide2 import QtCore, QtWidgets
-from modules.Form import Ui_Auxprog
+from modules.Form import Ui_Form
 
-class Main(Ui_Auxprog, QtWidgets.QWidget):
+class Main(Ui_Form, QtWidgets.QWidget):
     """Код автокликера"""
     def __init__(self):
         super().__init__()
@@ -15,14 +15,9 @@ class Main(Ui_Auxprog, QtWidgets.QWidget):
         #Сигналы-слоты
         self.timer = QtCore.QTimer(self)
         self.timer.setInterval(self.tempo)
-        self.Delay01.editingFinished.connect(self.set_tempo)
         self.timer.timeout.connect(self.auto_click)
         self.click_status = False
         self.Startstop.clicked.connect(self.button_click)
-
-    def set_tempo(self):
-        self.tempo = int(self.Delay01.text())
-        print(self.tempo)
 
     def button_click(self):
         """Автокликер-Переключатель"""
@@ -37,15 +32,8 @@ class Main(Ui_Auxprog, QtWidgets.QWidget):
     def auto_click(self):
         """Автокликек-Кликер"""
         a.click()
-        print('1')
 
-    def mouseMoveEvent(self, e):
+    def mousePressEvent(self, e):
         """Автокликер-Позиция курсора(Qt)"""
         self.posX.setText(str(e.x()))
         self.posY.setText(str(e.y()))
-
-    def keyPressEvent(self, e):
-        """Автокликер-Горячие клавиши(Qt)"""
-        if e.key():
-            #self.label_2.setText(str(chr(e.key())))
-        e.accept()
