+++ requirements.txt
@@ -1,0 +1,20 @@
+altgraph==0.17
+autogui==0.1.8
+future==0.18.2
+MouseInfo==0.1.3
+pefile==2019.4.18
+Pillow==8.1.0
+PyAutoGUI==0.9.52
+pycparser==2.20
+PyGetWindow==0.0.9
+pyinstaller==4.1
+pyinstaller-hooks-contrib==2020.11
+PyMsgBox==1.0.9
+pyperclip==1.8.1
+PyRect==0.1.4
+PyScreeze==0.1.26
+PySide2==5.15.2
+pythonnet==2.5.1
+PyTweening==1.0.3
+pywin32-ctypes==0.2.0
+shiboken2==5.15.2
